+++ src/components/Header/Header.jsx
@@ -1,11 +1,13 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import * as S from './style';
+import useModal from '../../hooks/useModal';
+import LoginModal from '../LoginModal/LoginModal';
 
 const Header = () => {
-  const [isLogin, setLogin] = useState(false);
   const navigate = useNavigate();
-
+  const [showModal, toggleModal] = useModal();
+  const [isLogin, setLogin] = useState(false);
   const goToPage = url => navigate(url);
 
   return (
@@ -27,8 +29,9 @@ const Header = () => {
         {isLogin === true ? (
           <S.UserImage />
         ) : (
-          <S.LoginButton>로그인</S.LoginButton>
+          <S.LoginButton onClick={toggleModal}>로그인</S.LoginButton>
         )}
+        {showModal && <LoginModal closeModal={toggleModal} />}
       </S.MenuBar>
     </S.Header>
   );
